Surface StorageProfile options
Follow-up to #2524

Make ApplyStorageProfileSettings component independent from
react-hook-form and re-use it for AddBootableVolumeModal and
CreateBootableVolumeModal. Both modals rely on useState hook to keep
the state.
To better encapsulate the functionality, the component will take over
the responsibility of clearing the volume/access modes after storage
class change. The storage class name is stored as ref to avoid using
setters as dependencies in the useEffect hook (component cannot assume
their references will be stable).

Backported from 4.19 #2653

Reference-Url: https://github.com/kubevirt-ui/kubevirt-plugin/pull/2653
Reference-Url: https://github.com/kubevirt-ui/kubevirt-plugin/pull/2524
