+++ locales/en/plugin__kubevirt-plugin.json
@@ -676,6 +676,7 @@
   "Learn how to create, import, and run virtual machines on OpenShift with step-by-step instructions and tasks.": "Learn how to create, import, and run virtual machines on OpenShift with step-by-step instructions and tasks.",
   "Learn how to use VirtualMachines": "Learn how to use VirtualMachines",
   "Learn more": "Learn more",
+  "Learn more about <2>StorageProfile</2>.": "Learn more about <2>StorageProfile</2>.",
   "Learn more about key areas to complete workflows, increase productivity, and familiarize yourself with Virtualization using our resources.": "Learn more about key areas to complete workflows, increase productivity, and familiarize yourself with Virtualization using our resources.",
   "Learn more about MigrationPolicies": "Learn more about MigrationPolicies",
   "Learn more about network latency checkups": "Learn more about network latency checkups",
@@ -686,7 +687,6 @@
   "Learn more about supported formats": "Learn more about supported formats",
   "Learn more about templates": "Learn more about templates",
   "Learn more about working with projects": "Learn more about working with projects",
-  "Learn more about<1> StorageProfile</1>.": "Learn more about<1> StorageProfile</1>.",
   "less than": "less than",
   "Limitations": "Limitations",
   "Linux only": "Linux only",
@@ -854,7 +854,6 @@
   "No network interfaces found": "No network interfaces found",
   "No network latency checkups found": "No network latency checkups found",
   "No NetworkAttachmentDefinitions available. Contact your system administrator for additional support.": "No NetworkAttachmentDefinitions available. Contact your system administrator for additional support.",
-  "No optimized StorageProfile settings for this StorageClass.": "No optimized StorageProfile settings for this StorageClass.",
   "No owner": "No owner",
   "No parameters found in this template.": "No parameters found in this template.",
   "No preferences found": "No preferences found",
@@ -921,7 +920,6 @@
   "Optimized for best performance. Supported by most Linux distributions. Windows requires additional drivers to use this model": "Optimized for best performance. Supported by most Linux distributions. Windows requires additional drivers to use this model",
   "Optimized for best performance. Supported by most Linux distributions. Windows requires additional drivers to use this model.": "Optimized for best performance. Supported by most Linux distributions. Windows requires additional drivers to use this model.",
   "Optimized for High resource consumption workloads": "Optimized for High resource consumption workloads",
-  "Optimized values Access mode: {{accessMode}}, Volume mode: {{volumeMode}}.": "Optimized values Access mode: {{accessMode}}, Volume mode: {{volumeMode}}.",
   "Optional parameters": "Optional parameters",
   "Organization ID": "Organization ID",
   "Other": "Other",
