+++ src/utils/components/ApplyStorageProfileSettings/ApplyStorageProfileSettings.scss
@@ -6,6 +6,6 @@
   &--labelWithGap {
     display: flex;
     gap: var(--pf-v5-global--spacer--sm);
-    align-items: anchor-center;
+    align-items: center;
   }
 }
