+++ src/utils/components/DiskModal/components/utils/modesMapping.ts
@@ -1,11 +1,7 @@
-import { uniq } from 'lodash';
-
 import {
   V1beta1StorageSpecAccessModesEnum,
   V1beta1StorageSpecVolumeModeEnum,
 } from '@kubevirt-ui/kubevirt-api/kubevirt';
-import { t } from '@kubevirt-utils/hooks/useKubevirtTranslation';
-import { ClaimPropertySets } from '@kubevirt-utils/types/storage';
 
 export const initialAccessModes: V1beta1StorageSpecAccessModesEnum[] = [
   V1beta1StorageSpecAccessModesEnum.ReadWriteMany,
@@ -200,32 +196,6 @@ export const provisionerAccessModeMapping: ProvisionerAccessModeMapping = {
   },
 };
 
-export const ACCESS_MODE_RADIO_OPTIONS = [
-  {
-    label: t('Single user (RWO)'),
-    value: V1beta1StorageSpecAccessModesEnum.ReadWriteOnce,
-  },
-  {
-    label: t('Shared access (RWX)'),
-    value: V1beta1StorageSpecAccessModesEnum.ReadWriteMany,
-  },
-  {
-    label: t('Read only (ROX)'),
-    value: V1beta1StorageSpecAccessModesEnum.ReadOnlyMany,
-  },
-];
-
-export const VOLUME_MODE_RADIO_OPTIONS = [
-  {
-    label: t('Filesystem'),
-    value: V1beta1StorageSpecVolumeModeEnum.Filesystem,
-  },
-  {
-    label: t('Block'),
-    value: V1beta1StorageSpecVolumeModeEnum.Block,
-  },
-];
-
 export const getAccessModeForProvisioner = (
   provisioner: string,
   volumeMode: V1beta1StorageSpecVolumeModeEnum,
@@ -255,16 +225,3 @@ export const getVolumeModeForProvisioner = (
 
   return initialVolumeModes;
 };
-
-export const getAccessModesForVolume = (
-  claimPropertySets: ClaimPropertySets,
-  volumeMode?: string,
-) =>
-  uniq(
-    claimPropertySets
-      .filter((it) => it.volumeMode === volumeMode)
-      .flatMap((it) => it.accessModes)
-      .filter(Boolean)
-      .map((mode) => V1beta1StorageSpecAccessModesEnum[mode])
-      .filter(Boolean),
-  );
